Add files via upload
Given an array of n elements and a number m. we need to find all distinct pairs existing in the array whose pair sum is divisible by the given number

m.

Print the total number of such pairs. Distinct pairs means (1,2) and (2,1) are the same, i.e., ordering of the pairs doesn't matter.